Fix profiles (#45)
Only a single profile can be set as default. It gets applied only when you don't specify the profile in Istio.spec.profile. However, the profile (either the one selected in Istio.spec.profile or the default profile) is always applied on top of the profile called "default", but this profile is always applied on top of the profile called "default", just as when using istioctl install.

Signed-off-by: Marko Lukša <mluksa@redhat.com>