OSSM-5393 Add product-specific configuration (#10)
This creates a new ossm/ directory that we can use to store everything
related to only downstream. It also adds a Makefile.vendor.mk to
overwrite some variables in the Makefile and changes a field in the
PROJECT file (for operator-sdk - this is the only place where we
might encounter legitimate conflicts). All other changes have been
generated.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>