Don't log error when namespace is being deleted
Previously, when the namespace referenced in IstioCni or IstioRevision.spec.namespace was being deleted, the operator log would show this as an error (with a stacktrace). Because of the requeues, the error was also logged many times.

Now, if the namespace is being deleted, the operator logs this at INFO level. It also doesn't requeue the object in this case.

Signed-off-by: Marko Lukša <mluksa@redhat.com>