Automate the bump of stable Istio releases
Modify the existing `update-istio.sh` script to also look for new stable
versions and update `versions.yaml` if any is found.

This automates the inclusion of new patch releases (currently done in a
nightly job), but does not remove anything. Removals should be manual.

Note that it only looks for new patch releases of *existing* versions in
`versions.yaml` file. It does not look for a new *minor* version, only
patch versions. New minor or major versions should still be added
manually.

Signed-off-by: Jonh Wendell <jwendell@redhat.com>