Display a friendly message for missing K8s cluster (#175)
Some of the Makefile commands require a running K8s cluster.
When these commands are issued without an active cluster, we
currently do not display a proper message, and users have to
figure it out manually. This PR fixes that by verifying access
to a running cluster and displaying a friendly message if not.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
Signed-off-by: bmangoen <bmangoen@redhat.com>