Skip reconcile of IstioRevision on owned resource status change
Some resources owned by IstioRevision have a status. When this status changes, the operator doesn't need to reconcile the IstioRevision, unless the status affects the status of the IstioRevision itself (as is the case with the istiod Deployment status).

For example, the HorizontalPodAutoscaler's status changes often, because it contains the current metrics related to the istiod pod(s). These changes shouldn't cause the IstioRevision object to be reconciled.

Signed-off-by: Marko Lukša <mluksa@redhat.com>