Ensure controller-gen make target honors CONTROLLER_TOOLS_VERSION (#298)
Previously, bumping `CONTROLLER_TOOLS_VERSION` didn't cause an existing `controller-gen` binary to be updated. This was because the `controller-gen` make target had a dependency on the `$(CONTROLLER_GEN)` target. Since `$(CONTROLLER_GEN)` points to `bin/controller-gen`, the make target isn't invoked when the file exists. Because of this, the `controller-gen --version` check in the `$(CONTROLLER_GEN)` target was never executed and thus the binary was never updated.

Now, the `controller-gen` target will always perform the version check.

Signed-off-by: Marko Lukša <mluksa@redhat.com>