Convert all values to pointers
This allows us to differentiate between unset values and values explicitly set to an empty value (e.g. "", 0, or false). Without this, users can't override the defaults set in e.g. profiles.

For example, `cni.chained` defaults to `true`. Previously, if a user set this value to `false` in the Istio resource, the operator would create the `IstioRevision` without this it, because it thought that the value wasn't set due to `false` being the zero value for booleans. The end result was that Istio CNI was deployed with `chained` set to `true`.

Now, all boolean, string, and int/float values are pointers. This allows the operator to know whether the user has set the value or not and, if set, carry the value to the IstioRevision resource.

Signed-off-by: Marko Lukša <mluksa@redhat.com>