Show name of active revision in Istio/RemoteIstio status (#350)
This helps users know the name of the currently active revision. Previously, the users had to just know that for the InPlace strategy, the name of the revision is the same as the name of the Istio/RemoteIstio resource, and for the RevisionBased strategy, the revision name is the Istio/RemoteIstio name + version.

Signed-off-by: Marko Lukša <mluksa@redhat.com>