Add Helm artifacts publish automation
When a release being created, the helm artifacts stored as an assets of
the release.
In order to make those artifacts available for "helm repo add" flow, an
"index.yaml" file under "gh-pages" branch needs to be updated.

Once updated, it will serve the helm deployment method with the new
artifacts version.

Add an automation flow to github actions that could be triggered and
will make the required updates.
At the end, a pull request with the changes will be created.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>