Simplify supportedversion.go and add unit tests (#391)
* Simplify supportedversion.go and add unit tests

These tests will not just ensure that the functions in supportedversion.go are okay, but also that the versions.yaml file is valid and is parsed successfully.

Signed-off-by: Marko Lukša <mluksa@redhat.com>

* Make GetVersionFromIstiod return a semver.Version

Signed-off-by: Marko Lukša <mluksa@redhat.com>

* Make getProxyVersion() return semver.Version

Signed-off-by: Marko Lukša <mluksa@redhat.com>

* Include commit hash in latest version in versions.yaml

Without the commit hash, the version number doesn't specify a single version, but a set of versions. For example, "1.24-alpha" is not just one single version, whereas "1.24-alpha.<hash>" is.

Signed-off-by: Marko Lukša <mluksa@redhat.com>

---------

Signed-off-by: Marko Lukša <mluksa@redhat.com>