Support metrics on dualStack clusters (#426)
Sail Operator uses `kube-rbac-proxy` to serve metrics securely.
It is configured to listen on `secure-listen-address` with an
`upstream` that points to the actual metrics server in the Sail
Operator listening on the loopback interface. Currently, the
`secure-listen-address` is configured with "0.0.0.0:8443" and this
PR updates it to ":8443" so that it can listen on both IPv4 and
IPv6 interfaces.

This PR also modifies the spec.ipFamilyPolicy of
sail-operator-metrics-service to PreferDualStack, so that it can
support all types of clusters.

Related to: https://github.com/istio-ecosystem/sail-operator/issues/372
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>