Fix undeploy when using Ambient mode
When make undeploy is executed in an Ambient deployment, it hangs.
This happens because the Sail Operator is deleted, but a finalizer
remains on the ztunnel resource. As the finalizer isn't removed, the
ztunnel CR and the associated CRD cannot be deleted.

Related to: https://github.com/istio-ecosystem/sail-operator/issues/500
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>