e2e tests for validating Ambient deployment (#541)
* e2e tests for validating Ambient deployment

This PR implements e2e tests for control plane and dataplane validation
of Ambient profile in the Sail Operator.

Related to: https://github.com/istio-ecosystem/sail-operator/issues/500
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

* Fix flaky error seen while running tests

As part of the test framework, when istio and istio-cni namespaces are
deleted, we were not waiting until the namespaces are deleted. Because
of this, we were seeing some race conditions in subsequent tests. This
PR fixes it.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

* Verify ztunnel listening sockets in the application pod ns

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

* Avoid platform while running the e2e tests

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

---------

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>