Add e2e test ensuring revisions in use by remotes remains
Adds an e2e test that ensures a revision that is in use by a remote cluster is not pruned even when it is no longer in use by the primary cluster. This is necessary because the Sail Operator has no way of knowing when a revision that manages a remote cluster is no longer in use because it does not have access to the remote cluster to check the workloads there.

Signed-off-by: Nick Fox <nfox@redhat.com>