Add `make cluster` command
Add the ability to create a cluster for testing by using `make cluster`.
Multiple clusters can be created by utilizing the `MULTICLUSTER=true`
option.
This allows developers to easily create a cluster to run manual testing
on.

Most of the code has been carved out of `tests/e2e/integ-suite-kind.sh`
which now relies on the cluster existing (the Makefile has been updated
to create it when running the tests).

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>