Add Openshift support for multicluster test suite
While executing the multicluster test flow on "Kind" environment, it
uses "MetalLB" for multicluster environment preparation.
When running multicluster test on Openshift and installing IstioCNI
plugin, it conflicts with "MetalLB" plugin.

In order to keep the test executiong flow on "Kind" environment and be
able to execute it on Openshift, use a condition for IstioCNI plugin
install to be deployed on Openshift only.

Signed-off-by: Maxim Babushkin <mbabushk@redhat.com>