Check IstioCNI status for IstioRevisions that depend on IstioCNI (#711)
* Check IstioCNI status for IstioRevisions that depend on IstioCNI

Previously, when an IstioRevision depended on IstioCNI, either because the user enabled the use of IstioCNI through values.pilot.cni.enabled or because the platform profile enables it, the IstioRevision resource would show as healthy even though workloads would fail because IstioCNI hasn't been deployed. This commit adds a new `DependenciesHealthy` condition to `IstioRevision` and makes `IstioRevision.status.state` and the `STATUS` column report the fact that IstioCNI is missing or unhealthy. This makes it easier for users to see that their setup is misconfigured.

Signed-off-by: Marko Lukša <mluksa@redhat.com>

* Add integration test

Signed-off-by: Marko Lukša <mluksa@redhat.com>

---------

Signed-off-by: Marko Lukša <mluksa@redhat.com>