Add label app.kubernetes.io/managed-by: sail-operator
There was previously no mention of the Sail operator on the resources it deploys. The `app.kubernetes.io/managed-by` label indicated that the resources were managed by Helm, but this is somewhat misleading. Even though resources are indeed rendered by Helm, this is only an implementation detail. The resources are in fact managed by the Sail operator, so the label should reflect this.

This label will also allow the operator to watch for changes only in resources that are managed by it and ignore all others.

Signed-off-by: Marko Lukša <mluksa@redhat.com>