Add PROFILE column (#755)
The purpose of this column is mostly to make it immediately clear when an Istio deployment is actually a remote or ambient installation. We previously had a TYPE column on IstioRevision, which indicated whether the revision was remote or not. The PROFILE column improves this, since it's also used for other current and future profiles.

Signed-off-by: Marko Lukša <mluksa@redhat.com>