Speed up docs tests by running on a single cluster (#855)
Refactor the tests to run all tags on a single cluster without
redeploying it.
After each tag the cluster is cleaned back to its original state.

Dual stack tests are still being run in their own cluster, as it needs
to be deployed in dual stack mode to support it.

Fixes: https://github.com/istio-ecosystem/sail-operator/issues/827

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>