Manual Cherry pick for multiple e2e improvements (#900)
* Add KEEP_ON_FAILURE flag to e2e kind tests (#832)

Using this flag allows debugging the failed test on the environment,
instead of destroying it.
The default behavior is still to clean up the cluster when failing.

Fixes: https://github.com/istio-ecosystem/sail-operator/issues/675

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Simplify logic when using `KEEP_ON_FAILURE` (#858)

When using `KEEP_ON_FAILURE=true` also set `SKIP_CLEANUP=true`. This
keeps the cluster running after the tests, without a need to defined our
own trap logic.

The downside is that the cluster will keep running even if the tests
passed, but seeing as using the `KEEP_ON_FAILURE` flag is excpected
mainly when there are failures, it seems that it's resonable to do.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Speed up docs tests by running on a single cluster (#855)

Refactor the tests to run all tags on a single cluster without
redeploying it.
After each tag the cluster is cleaned back to its original state.

Dual stack tests are still being run in their own cluster, as it needs
to be deployed in dual stack mode to support it.

Fixes: https://github.com/istio-ecosystem/sail-operator/issues/827

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Improve e2e test cleanup by recording & cleaning up (#889)

* Consolidate namespace cleanup in a new package

The new package tracks namespaces and cleans up any namespaces it
doesn't know about.
The new `Cleaner` type can record and clean up the resources, given
it's called with `Record()` before the tests and `Cleanup()` after.

This is an initial change and more resources will be added.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Add CRD/CR cleanup logic for tests

Now the `Cleaner` can also record and clean up any rogue CRDs/CRs that
were created during the tests.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Replace operator uninstall with Cleanup

Now that Cleanup takes care of all resources that we care about, replace
the operator uninstall with it.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Add docs how to use `Cleaner`

Add docs to E2E tests about the `Cleaner` and how it should be used.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Add `make cluster` command (#705)

Add the ability to create a cluster for testing by using `make cluster`.
Multiple clusters can be created by utilizing the `MULTICLUSTER=true`
option.
This allows developers to easily create a cluster to run manual testing
on.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Co-authored-by: Mike Kolesnik <mkolesni@redhat.com>