fix IstioRevisionTag uninstall when revision namespace changes (#945)
Fixes #917.

This is quite a complex edge case, I'll do my best to explain. If you
look at the integration test I wrote (part of this commit), you can
follow along.

When a tag was created that referenced a revision in ns1, then patched
to reference a revision in ns2, we failed to clean up the old helm
install in ns1. That almost never causes problems - except in the case
where afterwards, you remove all tags, and then create a new one with
the same name that is never reconciled. If that tag is deleted, the operator will
attempt to uninstall that lingering helm release, which fails, leading
the operator to deadlock, and stopping the user from deleting the tag.
The only way to get the operator back into a working state would be to
manually delete that lingering helm release. That is a pretty
significant bug in my opinion.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>
Co-authored-by: Daniel Grimm <dgrimm@redhat.com>