Container: only use required packages (#949) (#1000)
This uses a two-stage build to produce a UBI-based image with only the
required packages, in this case only the setup package. buildinfo is
copied so that Clair has enough information to accurately scan the
image for vulnerable UBI packages.

The result is a minimal image with only the operator as the attack
surface (no shell etc.).

Signed-off-by: Stephen Kitt <skitt@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>
Co-authored-by: Stephen Kitt <steve@sk2.org>