Add base chart support (#802) (#994)
* Implement istiobase controller

The istiobase controller manages the application of the base chart. As only one instance of this chart can be installed at once, this has to be done by a new controller. If we had added the rendering of this chart to the istiorevision controller, we could have two instances of that controller both trying to update the chart/release.




* Add tests




* Handle case where IstioRevisionTag is modified to point to a revision in a different namespace




* lint watches in istiobase controller



* add permissions for managing ServiceAccounts



* Simplify test cleanup



* Remove istiobase controller



* install base chart into operator namespace



* fix test setup/cleanup



* populate OperatorNamespace field



* fix linter errors



* remove mention of istiobase controller from tests



---------

Signed-off-by: Marko Lukša <mluksa@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>
Co-authored-by: Marko Lukša <mluksa@redhat.com>