Revert default MaxConcurrentReconciles to 1 (#1167)
We increased this to 5 recently but we decided that we want more testing
around the feature before we change the default. Users can still
manually increase it.

I'm also reverting it in the tests to make sure the tests use the same
settings as a default deployment. We'll need to invest in proper testing
of different values.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>