fix helm linter by adjusting schema.json files
The gateway chart is incompatible with the new JSON schema linter used
in the latest upstream builder image. It'll be fixed in 1.27.1, but we
need to go back and adjust all older schema.json files. In fact, in the
latest `master` charts it's already fixed.

Sadly, this makes some unrelated changes to the formatting, but I prefer
that over hacking something together using `sed`.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>