Fix api_transformer's handling of inline comments (#1195)
Copying inline comments leads to corruption of the output sometimes. I
still don't fully understand how - it's likely because of how we're
re-using existing AST nodes to create the new AST and don't recreate the
position information in it.

This fixes the problems we're seeing on release-1.27 branch when adding
the NativeNftables field.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>
Co-authored-by: Daniel Grimm <dgrimm@redhat.com>