fix: remove MeshConfig from ZTunnel CRD
This removes the MeshConfig field from the ZTunnel CRD, which was not operational anyway.
Only one field from MeshConfig was ever used, defaultConfig.proxyMetadata, the content
of which was written verbatim to the DaemonSet's environment variables. We'll be able to
set this field automatically once we introduce `targetRef` support, and until then,
should users require custom values, they can use `ztunnel.env` which has the same effect.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>