Adding CI env var to e2e test workflow
This variable controls the creation of the internal image registry route for OCP cluster while running the test. This means that we need to set the creation of the route during a separated ci steps, this will add stability and control for the test workflow. Also the env var can be used to set others CI specific features later on

Signed-off-by: Francisco Herrera <fjglira@gmail.com>