[release-1.27] Enhance IstioRevision dependency check to include ztunnel status (#1383)
* Enhance IstioRevision dependency check to include ztunnel status

The IstioRevision controller was missing health validation for
ZTunnel while using the ambient mode. Because of this, even when
ztunnel was missing/unhealthy, the IstioRevision was showing the
status as Ready. This PR fixes it.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

* Add files from make gen

This commit includes changes in sailoperator.io.md file

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

* Add integration tests

This PR includes an integration test to validate the changes.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>

---------

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>
Co-authored-by: Sridhar Gaddam <sgaddam@redhat.com>