E2E code consolidations (#1397)
* E2E: Extract common function to await conditions

This code is very repetitive and verbose, extracting it to a common
function allows the tests to be much more concise and readable, without
affecting functionality.
The main benefit of this is making the tests easier to read and
maintain, and avoid repeating code (and making mistakes).

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* E2E: Add specific AwaitDeployment function

Most of the tests that await for the DeploymentAvailable condition do it
on the control plane namespace for a Deployment.
Extracting this specific case makes the tests even easier to read and
understand.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* E2E: Add CheckSamplePodsReady function

Reuse CheckPodsReady and add a specific function CheckSamplePodsReady
since most of the calls are on the sample namespace.
Both checks are now in the `checks.go` file for logical consistency.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* E2E: Extract waiting for CNI DaemonSet to be ready

This is repetitive and verbose, extracting this makes the tests easier to
read and maintain.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>